requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
aws-cdk-lib==2.70.0
aws-cdk.asset-awscli-v1==2.2.121
aws-cdk.asset-kubectl-v20==2.1.1
aws-cdk.asset-node-proxy-agent-v5==2.0.98
black==23.3.0
bleach==6.0.0
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-log==0.4.0
colorama==0.4.6
constructs==10.1.293
cryptography==39.0.2
dataclasses-json==0.5.7
deprecation==2.1.0
docutils==0.19
dotty-dict==1.3.1
exceptiongroup==1.1.1
fastapi==0.95.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
gotrue==1.0.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
invoke==1.7.3
jaraco.classes==3.2.3
jeepney==0.8.0
jsii==1.79.0
keyring==23.13.1
langchain==0.0.127
llama-index==0.5.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
more-itertools==9.1.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
openai==0.27.2
packaging==23.0
pandas==1.5.3
pathlib==1.0.1
pathspec==0.11.1
pkginfo==1.9.6
platformdirs==3.2.0
postgrest==0.10.6
publication==0.0.3
pycparser==2.21
pydantic==1.10.6
Pygments==2.14.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gitlab==3.13.0
python-multipart==0.0.6
python-semantic-release==7.33.2
pytz==2022.7.1
PyYAML==6.0
readme-renderer==37.3
realtime==1.0.0
regex==2022.10.31
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==1.5.0
SecretStorage==3.3.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==1.4.47
sqlitedict==2.1.0
starlette==0.26.1
storage3==0.5.2
StrEnum==0.4.10
supabase==1.0.2
supafunc==0.2.2
tenacity==8.2.2
tiktoken==0.3.3
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.65.0
twine==3.8.0
typeguard==2.13.3
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.20.0
webencodings==0.5.1
websockets==10.4
yarl==1.8.2
zipp==3.15.0
